LPS-74409 More accurate name and package. Renamed both PortletDisplayTemplateDDMPermissionSupport and AssetDisplayTemplateDDMPermissionSupport to make it *DDMTemplatePermissionSupport. And moved AssetDisplayTemplateDDMTemplatePermissionSupport to match the package in PortletDisplayTemplateDDMTemplatePermissionSupport.
