+++ modules/apps/web-experience/portlet-display-template/portlet-display-template-web/src/main/java/com/liferay/portlet/display/template/web/internal/dynamic/data/mapping/util/PortletDisplayTemplateDDMTemplatePermissionSupport.java
@@ -33,7 +33,7 @@ import org.osgi.service.component.annotations.Component;
 		"model.class.name=com.liferay.portlet.display.template.PortletDisplayTemplate"
 	}
 )
-public class PortletDisplayTemplateDDMPermissionSupport
+public class PortletDisplayTemplateDDMTemplatePermissionSupport
 	implements DDMTemplatePermissionSupport {
 
 	@Override
